Decoupling of platform-agnostic and platform-specific SessionOptions functionality
