+++ csharp/src/Microsoft.ML.OnnxRuntime/NativeLib.netcore.cs
@@ -1,0 +1,7 @@
+﻿namespace Microsoft.ML.OnnxRuntime
+{
+    internal class NativeLib
+    {
+        internal const string DllName = "onnxruntime";
+    }
+}
