+++ csharp/src/Microsoft.ML.OnnxRuntime/NativeMethods.shared.cs
@@ -229,7 +229,7 @@ namespace Microsoft.ML.OnnxRuntime
     {
         static OrtApi api_;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate ref OrtApi DOrtGetApi(UInt32 version);
@@ -391,26 +391,26 @@ namespace Microsoft.ML.OnnxRuntime
 
         #region Runtime/Environment API
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus* */DOrtCreateEnv(LogLevel default_warning_level, string logId, out IntPtr /*(OrtEnv*)*/ env);
         public static DOrtCreateEnv OrtCreateEnv;
 
         // OrtReleaseEnv should not be used
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtReleaseEnv(IntPtr /*(OrtEnv*)*/ env);
         public static DOrtReleaseEnv OrtReleaseEnv;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus* */DOrtEnableTelemetryEvents(IntPtr /*(OrtEnv*)*/ env);
         public static DOrtEnableTelemetryEvents OrtEnableTelemetryEvents;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus* */DOrtDisableTelemetryEvents(IntPtr /*(OrtEnv*)*/ env);
@@ -424,7 +424,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// Creates native OrtTensorRTProviderOptions instance
         /// </summary>
         /// <param name="trtProviderOptionsInstance">(output) native instance of OrtTensorRTProviderOptions</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus* */DOrtCreateTensorRTProviderOptions(
@@ -438,7 +438,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="providerOptionsKeys">configuration keys of OrtTensorRTProviderOptions</param>
         /// <param name="providerOptionsValues">configuration values of OrtTensorRTProviderOptions</param>
         /// <param name="numKeys">number of configuration keys</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus* */DOrtUpdateTensorRTProviderOptions(
@@ -453,7 +453,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// </summary>
         /// <param name="allocator">instance of OrtAllocator</param>
         /// <param name="ptr">is a UTF-8 null terminated string allocated using 'allocator'</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus* */DOrtGetTensorRTProviderOptionsAsString(
@@ -466,7 +466,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// Releases native OrtTensorRTProviderOptions instance
         /// </summary>
         /// <param name="trtProviderOptionsInstance">native instance of OrtTensorRTProviderOptions to be released</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtReleaseTensorRTProviderOptions(IntPtr /*(OrtTensorRTProviderOptions*)*/ trtProviderOptionsInstance);
@@ -475,7 +475,7 @@ namespace Microsoft.ML.OnnxRuntime
         #endregion
 
         #region Status API
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate ErrorCode DOrtGetErrorCode(IntPtr /*(OrtStatus*)*/status);
@@ -483,13 +483,13 @@ namespace Microsoft.ML.OnnxRuntime
 
         // returns char*, need to convert to string by the caller.
         // does not free the underlying OrtStatus*
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* char* */DOrtGetErrorMessage(IntPtr /* (OrtStatus*) */status);
         public static DOrtGetErrorMessage OrtGetErrorMessage;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtReleaseStatus(IntPtr /*(OrtStatus*)*/ statusPtr);
@@ -499,7 +499,7 @@ namespace Microsoft.ML.OnnxRuntime
 
         #region InferenceSession API
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus* */DOrtCreateSession(
@@ -518,7 +518,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="sessionOptions">Native SessionOptions instance</param>         
         /// <param name="prepackedWeightsContainer">Native OrtPrepackedWeightsContainer instance</param>
         /// <param name="session">(Output) Created native OrtSession instance</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus* */DOrtCreateSessionWithPrepackedWeightsContainer(
@@ -529,7 +529,7 @@ namespace Microsoft.ML.OnnxRuntime
                                         out IntPtr /* (OrtSession**) */ session);
         public static DOrtCreateSessionWithPrepackedWeightsContainer OrtCreateSessionWithPrepackedWeightsContainer;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus* */DOrtCreateSessionFromArray(
@@ -549,7 +549,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="sessionOptions">Native SessionOptions instance</param>         
         /// <param name="prepackedWeightsContainer">Native OrtPrepackedWeightsContainer instance</param>
         /// <param name="session">(Output) Created native OrtSession instance</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus* */DOrtCreateSessionFromArrayWithPrepackedWeightsContainer(
@@ -561,7 +561,7 @@ namespace Microsoft.ML.OnnxRuntime
                                         out IntPtr /* (OrtSession**) */ session);
         public static DOrtCreateSessionFromArrayWithPrepackedWeightsContainer OrtCreateSessionFromArrayWithPrepackedWeightsContainer;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(ONNStatus*)*/ DOrtRun(
@@ -576,7 +576,7 @@ namespace Microsoft.ML.OnnxRuntime
                                                 );
         public static DOrtRun OrtRun;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(ONNStatus*)*/ DOrtRunWithBinding(
@@ -586,7 +586,7 @@ namespace Microsoft.ML.OnnxRuntime
                                                 );
         public static DOrtRunWithBinding OrtRunWithBinding;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtSessionGetInputCount(
@@ -594,7 +594,7 @@ namespace Microsoft.ML.OnnxRuntime
                                                 out UIntPtr count);
         public static DOrtSessionGetInputCount OrtSessionGetInputCount;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtSessionGetOutputCount(
@@ -602,7 +602,7 @@ namespace Microsoft.ML.OnnxRuntime
                                                 out UIntPtr count);
         public static DOrtSessionGetOutputCount OrtSessionGetOutputCount;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtSessionGetOverridableInitializerCount(
@@ -610,7 +610,7 @@ namespace Microsoft.ML.OnnxRuntime
                                                 out UIntPtr count);
         public static DOrtSessionGetOverridableInitializerCount OrtSessionGetOverridableInitializerCount;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/DOrtSessionGetInputName(
@@ -620,7 +620,7 @@ namespace Microsoft.ML.OnnxRuntime
                                                 out IntPtr /*(char**)*/name);
         public static DOrtSessionGetInputName OrtSessionGetInputName;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/DOrtSessionGetOutputName(
@@ -630,7 +630,7 @@ namespace Microsoft.ML.OnnxRuntime
                                                 out IntPtr /*(char**)*/name);
         public static DOrtSessionGetOutputName OrtSessionGetOutputName;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/DOrtSessionEndProfiling(
@@ -639,7 +639,7 @@ namespace Microsoft.ML.OnnxRuntime
                                                 out IntPtr /*(char**)*/profile_file);
         public static DOrtSessionEndProfiling OrtSessionEndProfiling;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/DOrtSessionGetOverridableInitializerName(
@@ -649,7 +649,7 @@ namespace Microsoft.ML.OnnxRuntime
                                                 out IntPtr /*(char**)*/name);
         public static DOrtSessionGetOverridableInitializerName OrtSessionGetOverridableInitializerName;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/DOrtSessionGetInputTypeInfo(
@@ -658,7 +658,7 @@ namespace Microsoft.ML.OnnxRuntime
                                                 out IntPtr /*(struct OrtTypeInfo**)*/ typeInfo);
         public static DOrtSessionGetInputTypeInfo OrtSessionGetInputTypeInfo;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/DOrtSessionGetOutputTypeInfo(
@@ -667,7 +667,7 @@ namespace Microsoft.ML.OnnxRuntime
                                                 out IntPtr /* (struct OrtTypeInfo**)*/ typeInfo);
         public static DOrtSessionGetOutputTypeInfo OrtSessionGetOutputTypeInfo;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/DOrtSessionGetOverridableInitializerTypeInfo(
@@ -677,19 +677,19 @@ namespace Microsoft.ML.OnnxRuntime
         public static DOrtSessionGetOverridableInitializerTypeInfo OrtSessionGetOverridableInitializerTypeInfo;
 
         // release the typeinfo using OrtReleaseTypeInfo
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtReleaseTypeInfo(IntPtr /*(OrtTypeInfo*)*/session);
         public static DOrtReleaseTypeInfo OrtReleaseTypeInfo;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtReleaseSession(IntPtr /*(OrtSession*)*/session);
         public static DOrtReleaseSession OrtReleaseSession;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtSessionGetProfilingStartTimeNs(
@@ -701,104 +701,104 @@ namespace Microsoft.ML.OnnxRuntime
 
         #region SessionOptions API
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtCreateSessionOptions(out IntPtr /*(OrtSessionOptions**)*/ sessionOptions);
         public static DOrtCreateSessionOptions OrtCreateSessionOptions;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtReleaseSessionOptions(IntPtr /*(OrtSessionOptions*)*/session);
         public static DOrtReleaseSessionOptions OrtReleaseSessionOptions;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtCloneSessionOptions(IntPtr /*(OrtSessionOptions*)*/ sessionOptions, out IntPtr /*(OrtSessionOptions**)*/ output);
         public static DOrtCloneSessionOptions OrtCloneSessionOptions;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtSetSessionExecutionMode(IntPtr /*(OrtSessionOptions*)*/ options,
         ExecutionMode execution_mode);
         public static DOrtSetSessionExecutionMode OrtSetSessionExecutionMode;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtSetOptimizedModelFilePath(IntPtr /* OrtSessionOptions* */ options, byte[] optimizedModelFilepath);
         public static DOrtSetOptimizedModelFilePath OrtSetOptimizedModelFilePath;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtEnableProfiling(IntPtr /* OrtSessionOptions* */ options, byte[] profilePathPrefix);
         public static DOrtEnableProfiling OrtEnableProfiling;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtDisableProfiling(IntPtr /* OrtSessionOptions* */ options);
         public static DOrtDisableProfiling OrtDisableProfiling;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtEnableMemPattern(IntPtr /* OrtSessionOptions* */ options);
         public static DOrtEnableMemPattern OrtEnableMemPattern;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtDisableMemPattern(IntPtr /* OrtSessionOptions* */ options);
         public static DOrtDisableMemPattern OrtDisableMemPattern;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtEnableCpuMemArena(IntPtr /* OrtSessionOptions* */ options);
         public static DOrtEnableCpuMemArena OrtEnableCpuMemArena;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtDisableCpuMemArena(IntPtr /* OrtSessionOptions* */ options);
         public static DOrtDisableCpuMemArena OrtDisableCpuMemArena;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtSetSessionLogId(IntPtr /* OrtSessionOptions* */ options, IntPtr /* const char* */logId);
         public static DOrtSetSessionLogId OrtSetSessionLogId;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtSetSessionLogVerbosityLevel(IntPtr /* OrtSessionOptions* */ options, int sessionLogVerbosityLevel);
         public static DOrtSetSessionLogVerbosityLevel OrtSetSessionLogVerbosityLevel;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtSetSessionLogSeverityLevel(IntPtr /* OrtSessionOptions* */ options, OrtLoggingLevel sessionLogSeverityLevel);
         public static DOrtSetSessionLogSeverityLevel OrtSetSessionLogSeverityLevel;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtSetIntraOpNumThreads(IntPtr /* OrtSessionOptions* */ options, int intraOpNumThreads);
         public static DOrtSetIntraOpNumThreads OrtSetIntraOpNumThreads;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtSetInterOpNumThreads(IntPtr /* OrtSessionOptions* */ options, int interOpNumThreads);
         public static DOrtSetInterOpNumThreads OrtSetInterOpNumThreads;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtSetSessionGraphOptimizationLevel(IntPtr /* OrtSessionOptions* */ options, GraphOptimizationLevel graphOptimizationLevel);
@@ -810,7 +810,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="options">Native SessionOptions instance</param>
         /// <param name="configKey">Config key</param>
         /// <param name="configValue">Config value</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtAddSessionConfigEntry(IntPtr /* OrtSessionOptions* */ options,
@@ -826,18 +826,6 @@ namespace Microsoft.ML.OnnxRuntime
         [DllImport(NativeLib.DllName, CharSet = CharSet.Ansi)]
         public static extern IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_CPU(IntPtr /*(OrtSessionOptions*) */ options, int use_arena);
 
-#if __IOS__
-        public static IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_Dnnl(IntPtr /*(OrtSessionOptions*) */ options, int use_arena) => throw new NotImplementedException();
-        public static IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_CUDA(IntPtr /*(OrtSessionOptions*) */ options, int device_id) => throw new NotImplementedException();
-        public static IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_ROCM(IntPtr /*(OrtSessionOptions*) */ options, int device_id) => throw new NotImplementedException();
-        public static IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_DML(IntPtr /*(OrtSessionOptions*) */ options, int device_id) => throw new NotImplementedException();
-        public static IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_OpenVINO(IntPtr /*(OrtSessionOptions*)*/ options, IntPtr /*(const char*)*/ device_id) => throw new NotImplementedException();
-        public static IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_Tensorrt(IntPtr /*(OrtSessionOptions*)*/ options, int device_id) => throw new NotImplementedException();
-        public static IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_MIGraphX(IntPtr /*(OrtSessionOptions*)*/ options, int device_id) => throw new NotImplementedException();
-        public static IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_Nnapi(IntPtr /*(OrtSessionOptions*)*/ options, uint nnapi_flags) => throw new NotImplementedException();
-        public static IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_Nuphar(IntPtr /*(OrtSessionOptions*) */ options, int allow_unaligned_buffers, IntPtr /*(char char*)*/ settings) => throw new NotImplementedException();
-        public static void OrtAddCustomOp(IntPtr /*(OrtSessionOptions*)*/ options, string custom_op_path) => throw new NotImplementedException();
-#else
         [DllImport(NativeLib.DllName, CharSet = CharSet.Ansi)]
         public static extern IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_Dnnl(IntPtr /*(OrtSessionOptions*) */ options, int use_arena);
 
@@ -845,15 +833,13 @@ namespace Microsoft.ML.OnnxRuntime
         public static extern IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_CUDA(IntPtr /*(OrtSessionOptions*) */ options, int device_id);
 
         [DllImport(NativeLib.DllName, CharSet = CharSet.Ansi)]
-        public static extern IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_ROCM(
-            IntPtr /*(OrtSessionOptions*) */ options, int device_id, UIntPtr gpu_mem_limit);
+        public static extern IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_ROCM(IntPtr /*(OrtSessionOptions*) */ options, int device_id, UIntPtr gpu_mem_limit);
 
         [DllImport(NativeLib.DllName, CharSet = CharSet.Ansi)]
         public static extern IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_DML(IntPtr /*(OrtSessionOptions*) */ options, int device_id);
 
         [DllImport(NativeLib.DllName, CharSet = CharSet.Ansi)]
-        public static extern IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_OpenVINO(
-            IntPtr /*(OrtSessionOptions*)*/ options, IntPtr /*(const char*)*/ device_id);
+        public static extern IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_OpenVINO(IntPtr /*(OrtSessionOptions*)*/ options, IntPtr /*(const char*)*/ device_id);
 
         [DllImport(NativeLib.DllName, CharSet = CharSet.Ansi)]
         public static extern IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_Tensorrt(IntPtr /*(OrtSessionOptions*)*/ options, int device_id);
@@ -865,20 +851,17 @@ namespace Microsoft.ML.OnnxRuntime
         public static extern IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_Nnapi(IntPtr /*(OrtSessionOptions*)*/ options, uint nnapi_flags);
 
         [DllImport(NativeLib.DllName, CharSet = CharSet.Ansi)]
-        public static extern IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_Nuphar(IntPtr /*(OrtSessionOptions*) */ options,
-                                                                                                     int allow_unaligned_buffers,
-                                                                                                     IntPtr /*(char char*)*/ settings);
+        public static extern IntPtr /*(OrtStatus*)*/ OrtSessionOptionsAppendExecutionProvider_Nuphar(IntPtr /*(OrtSessionOptions*) */ options, int allow_unaligned_buffers, IntPtr /*(char char*)*/ settings);
 
         [DllImport(NativeLib.DllName, CharSet = CharSet.Ansi)]
         public static extern void OrtAddCustomOp(IntPtr /*(OrtSessionOptions*)*/ options, string custom_op_path);
-#endif
 
         /// <summary>
         /// Append a TensorRT EP instance (configured based on given provider options) to the native OrtSessionOptions instance
         /// </summary>
         /// <param name="options">Native OrtSessionOptions instance</param>
         /// <param name="trtProviderOptions">Native OrtTensorRTProviderOptions instance</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/DSessionOptionsAppendExecutionProvider_TensorRT(
@@ -892,7 +875,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// </summary>
         /// <param name="options">Native OrtSessionOptions instance</param>
         /// <param name="trtProviderOptions">Native OrtTensorRTProviderOptionsV2 instance</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/DSessionOptionsAppendExecutionProvider_TensorRT_V2(
@@ -907,7 +890,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="options">Native SessionOptions instance</param>
         /// <param name="dimDenotation">Dimension denotation</param>
         /// <param name="dimValue">Dimension value</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/DOrtAddFreeDimensionOverride(IntPtr /*(OrtSessionOptions*)*/ options,
@@ -922,7 +905,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="options">Native SessionOptions instance</param>
         /// <param name="dimName">Dimension name</param>
         /// <param name="dimValue">Dimension value</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/DOrtAddFreeDimensionOverrideByName(IntPtr /*(OrtSessionOptions*)*/ options,
@@ -937,7 +920,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="options">Native SessionOptions instance</param>
         /// <param name="libraryPath">Library path</param>
         /// <param name="libraryHandle">(out) Native library handle</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/DOrtRegisterCustomOpsLibrary(IntPtr /*(OrtSessionOptions*) */ options,
@@ -952,7 +935,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="options">Native SessionOptions instance</param>
         /// <param name="name">Name of the initializer</param>
         /// <param name="ortValue">Native OrtValue instnce</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/DOrtAddInitializer(IntPtr /*(OrtSessionOptions*)*/ options,
@@ -965,49 +948,49 @@ namespace Microsoft.ML.OnnxRuntime
 
         #region RunOptions API
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtCreateRunOptions(out IntPtr /* OrtRunOptions** */ runOptions);
         public static DOrtCreateRunOptions OrtCreateRunOptions;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtReleaseRunOptions(IntPtr /*(OrtRunOptions*)*/options);
         public static DOrtReleaseRunOptions OrtReleaseRunOptions;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtRunOptionsSetRunLogVerbosityLevel(IntPtr /* OrtRunOptions* */ options, int value);
         public static DOrtRunOptionsSetRunLogVerbosityLevel OrtRunOptionsSetRunLogVerbosityLevel;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtRunOptionsSetRunLogSeverityLevel(IntPtr /* OrtRunOptions* */ options, OrtLoggingLevel value);
         public static DOrtRunOptionsSetRunLogSeverityLevel OrtRunOptionsSetRunLogSeverityLevel;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtRunOptionsSetRunTag(IntPtr /* OrtRunOptions* */ options, IntPtr /* const char* */ runTag);
         public static DOrtRunOptionsSetRunTag OrtRunOptionsSetRunTag;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtRunOptionsGetRunLogVerbosityLevel(IntPtr /* OrtRunOptions* */ options, out int verbosityLevel);
         public static DOrtRunOptionsGetRunLogVerbosityLevel OrtRunOptionsGetRunLogVerbosityLevel;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtRunOptionsGetRunLogSeverityLevel(IntPtr /* OrtRunOptions* */ options, out OrtLoggingLevel severityLevel);
         public static DOrtRunOptionsGetRunLogSeverityLevel OrtRunOptionsGetRunLogSeverityLevel;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtRunOptionsGetRunTag(IntPtr /* const OrtRunOptions* */options, out IntPtr /* const char** */ runtag);
@@ -1015,13 +998,13 @@ namespace Microsoft.ML.OnnxRuntime
 
         // Set a flag so that any running OrtRun* calls that are using this instance of OrtRunOptions
         // will exit as soon as possible if the flag is true.
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtRunOptionsSetTerminate(IntPtr /* OrtRunOptions* */ options);
         public static DOrtRunOptionsSetTerminate OrtRunOptionsSetTerminate;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtRunOptionsUnsetTerminate(IntPtr /* OrtRunOptions* */ options);
@@ -1031,7 +1014,7 @@ namespace Microsoft.ML.OnnxRuntime
 
         #region Allocator/MemoryInfo API
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* (OrtStatus*)*/ DOrtCreateMemoryInfo(
@@ -1044,7 +1027,7 @@ namespace Microsoft.ML.OnnxRuntime
 
         public static DOrtCreateMemoryInfo OrtCreateMemoryInfo;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* (OrtStatus*)*/ DOrtCreateCpuMemoryInfo(
@@ -1055,14 +1038,14 @@ namespace Microsoft.ML.OnnxRuntime
 
         public static DOrtCreateCpuMemoryInfo OrtCreateCpuMemoryInfo;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtReleaseMemoryInfo(IntPtr /*(OrtMemoryInfo*)*/ allocatorInfo);
 
         public static DOrtReleaseMemoryInfo OrtReleaseMemoryInfo;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtCompareMemoryInfo(
@@ -1075,21 +1058,21 @@ namespace Microsoft.ML.OnnxRuntime
         /**
         * Do not free the returned value
         */
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtMemoryInfoGetName(IntPtr /*(const OrtMemoryInfo* ptr)*/ mem_info, out IntPtr /*(const char**)*/ name);
 
         public static DOrtMemoryInfoGetName OrtMemoryInfoGetName;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtMemoryInfoGetId(IntPtr /*(const OrtMemoryInfo* ptr)*/ mem_info, out int /*(int* out)*/ id);
 
         public static DOrtMemoryInfoGetId OrtMemoryInfoGetId;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtMemoryInfoGetMemType(
@@ -1098,7 +1081,7 @@ namespace Microsoft.ML.OnnxRuntime
 
         public static DOrtMemoryInfoGetMemType OrtMemoryInfoGetMemType;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtMemoryInfoGetType(
@@ -1108,14 +1091,14 @@ namespace Microsoft.ML.OnnxRuntime
 
         public static DOrtMemoryInfoGetType OrtMemoryInfoGetType;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/DOrtGetAllocatorWithDefaultOptions(out IntPtr /*(OrtAllocator**)*/ allocator);
 
         public static DOrtGetAllocatorWithDefaultOptions OrtGetAllocatorWithDefaultOptions;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/DOrtAllocatorGetInfo(IntPtr /*(const OrtAllocator*)*/ ptr, out IntPtr /*(const struct OrtMemoryInfo**)*/info);
@@ -1131,7 +1114,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="initialChunkSizeBytes">Size of the region that the arena allocates first</param>
         /// <param name="maxDeadBytesPerChunk">Maximum amount of fragmentation allowed per chunk</param>
         /// <returns>Pointer to a native OrtStatus instance indicating success/failure of config creation</returns>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtCreateArenaCfg(UIntPtr /*(size_t)*/ maxMemory, int /*(int)*/ arenaExtendStrategy,
@@ -1144,7 +1127,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// Destroy an instance of an arena configuration instance
         /// </summary>
         /// <param name="arenaCfg">arena configuration instance to be destroyed</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtReleaseArenaCfg(IntPtr /*(OrtArenaCfg*)*/ arenaCfg);
@@ -1157,7 +1140,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="session">Session that this allocator should be used with</param>
         /// <param name="info">memory allocator specs</param>
         /// <param name="allocator">out pointer to a new allocator instance</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtCreateAllocator(IntPtr /*(const OrtSession*)*/ session, IntPtr /*(const OrtMemoryInfo*)*/ info, out IntPtr /*(OrtAllocator**)*/ allocator);
@@ -1168,7 +1151,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// Destroy an instance of an allocator created by OrtCreateAllocator
         /// </summary>
         /// <param name="allocator">instance to be destroyed</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtReleaseAllocator(IntPtr /*(OrtAllocator*)*/ allocator);
@@ -1181,7 +1164,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="allocator">allocator instance</param>
         /// <param name="size">bytes to allocate</param>
         /// <param name="p">out pointer to the allocated memory. Must be freed by OrtAllocatorFree</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr DOrtAllocatorAlloc(IntPtr /*(OrtAllocator*)*/ allocator, UIntPtr /*size_t*/ size, out IntPtr /*(void**)*/ p);
@@ -1193,7 +1176,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// </summary>
         /// <param name="allocator">allocator instance</param>
         /// <param name="p">pointer to native memory allocated by the allocator instance</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr DOrtAllocatorFree(IntPtr /*(OrtAllocator*)*/ allocator, IntPtr /*(void*)*/ p);
@@ -1212,7 +1195,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// </summary>
         /// <param name="session">session to create OrtIoBinding instance</param>
         /// <param name="io_binding">out a new instance of OrtIoBinding</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus*/ DOrtCreateIoBinding(IntPtr /*(const OrtSession*)*/ session, out IntPtr /*(OrtIoBinding)*/ io_binding);
@@ -1223,7 +1206,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// Destroy OrtIoBinding instance created by OrtCreateIoBinding
         /// </summary>
         /// <param name="io_bidning">instance of OrtIoBinding</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtReleaseIoBinding(IntPtr /*(OrtIoBinding)*/ io_binding);
@@ -1239,7 +1222,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="ort_value">OrtValue that is used for input (may wrap arbitrary memory).
         ///      The param instance is copied internally so this argument may be released.
         /// </param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus*/ DOrtBindInput(IntPtr /*(OrtIoBinding)*/ io_binding, IntPtr /*(const char*)*/ name, IntPtr /*const OrtValue**/ ort_value);
@@ -1255,7 +1238,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="ort_value">OrtValue that is used for output (may wrap arbitrary memory).
         ///      The param instance is copied internally so this argument may be released.
         /// </param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus*/ DOrtBindOutput(IntPtr /*(OrtIoBinding)*/ io_binding, IntPtr /*(const char*) */ name, IntPtr /*const OrtValue**/ ort_value);
@@ -1271,7 +1254,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="name">UTF-8 zero terminated name</param>
         /// <param name="mem_info">OrtMemoryInfo instance that contains device id. May be obtained from the device specific allocator instance</param>
         /// <returns></returns>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus*/ DOrtBindOutputToDevice(IntPtr /*(OrtIoBinding)*/ io_binding, IntPtr /*(const char*) */ name, IntPtr /* const OrtMemoryInfo */ mem_info);
@@ -1293,7 +1276,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// The buffer must be deallocated.</param>
         /// <param name="count">this contains the count of names returned which is the number of elements in lengths.</param>
         /// <returns></returns>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus*/ DOrtGetBoundOutputNames(IntPtr /* (const OrtIoBinding*) */ io_binding, IntPtr /* OrtAllocator* */ allocator,
@@ -1314,7 +1297,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="ortvalues">allocated buffer that contains pointers (IntPtr) to individual OrtValue instances</param>
         /// <param name="count">count of OrtValues returned</param>
         /// <returns></returns>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus*/ DOrtGetBoundOutputValues(IntPtr /* (const OrtIoBinding*) */ io_binding, IntPtr /* OrtAllocator* */ allocator,
@@ -1327,7 +1310,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// Releasing OrtIoBinding instance would clear all bound inputs.
         /// </summary>
         /// <param name="io_binding">instance of OrtIoBinding</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtClearBoundInputs(IntPtr /*(OrtIoBinding)*/ io_binding);
@@ -1339,7 +1322,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// Releasing OrtIoBinding instance would clear all bound outputs.
         /// </summary>
         /// <param name="io_binding">instance of OrtIoBinding</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtClearBoundOutputs(IntPtr /*(OrtIoBinding)*/ io_binding);
@@ -1352,7 +1335,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="location_values">a pointer to an array of index values that specify an element's location in the tensor data blob</param>
         /// <param name="location_values_count">length of location_values</param>
         /// <param name="out">a pointer to the element specified by location_values</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtTensorAt(IntPtr /*(OrtIoBinding)*/ io_binding);
@@ -1368,7 +1351,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="memInfo">Native OrtMemoryInfo instance</param>
         /// <param name="arenaCfg">Native OrtArenaCfg instance</param>
         /// <retruns>A pointer to native ortStatus indicating success/failure</retruns>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtCreateAndRegisterAllocator(IntPtr /*(OrtEnv*)*/ env,
@@ -1381,7 +1364,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// Set the language projection for collecting telemetry data when Env is created
         /// </summary>
         /// <param name="projection">the source projected language</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtSetLanguageProjection(IntPtr /* (OrtEnv*) */ environment, OrtLanguageProjection projection);
@@ -1397,7 +1380,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// </summary>
         /// <param name="session">instance of OrtSession</param>
         /// <param name="modelMetadata">(output) instance of OrtModelMetadata</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* (OrtStatus*) */ DOrtSessionGetModelMetadata(IntPtr /* (const OrtSession*) */ session, out IntPtr /* (OrtModelMetadata**) */ modelMetadata);
@@ -1410,7 +1393,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="modelMetadata">instance of OrtModelMetadata</param>
         /// <param name="allocator">instance of OrtAllocator</param>
         /// <param name="value">(output) producer name from the ModelMetadata instance</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* (OrtStatus*) */ DOrtModelMetadataGetProducerName(IntPtr /* (const OrtModelMetadata*) */ modelMetadata,
@@ -1424,7 +1407,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="modelMetadata">instance of OrtModelMetadata</param>
         /// <param name="allocator">instance of OrtAllocator</param>
         /// <param name="value">(output) graph name from the ModelMetadata instance</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* (OrtStatus*) */ DOrtModelMetadataGetGraphName(IntPtr /* (const OrtModelMetadata*) */ modelMetadata,
@@ -1438,7 +1421,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="modelMetadata">instance of OrtModelMetadata</param>
         /// <param name="allocator">instance of OrtAllocator</param>
         /// <param name="value">(output) domain from the ModelMetadata instance</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* (OrtStatus*) */ DOrtModelMetadataGetDomain(IntPtr /* (const OrtModelMetadata*) */ modelMetadata,
@@ -1453,7 +1436,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="allocator">instance of OrtAllocator</param>
         /// <param name="value">(output) description from the ModelMetadata instance</param>
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* (OrtStatus*) */ DOrtModelMetadataGetDescription(IntPtr /* (const OrtModelMetadata*) */ modelMetadata,
@@ -1467,7 +1450,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="modelMetadata">instance of OrtModelMetadata</param>
         /// <param name="allocator">instance of OrtAllocator</param>
         /// <param name="value">(output) graph description from the ModelMetadata instance</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* (OrtStatus*) */ DOrtModelMetadataGetGraphDescription(IntPtr /* (const OrtModelMetadata*) */ modelMetadata,
@@ -1480,7 +1463,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// </summary>
         /// <param name="modelMetadata">instance of OrtModelMetadata</param>
         /// <param name="value">(output) version from the ModelMetadata instance</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* (OrtStatus*) */ DOrtModelMetadataGetVersion(IntPtr /* (const OrtModelMetadata*) */ modelMetadata,
@@ -1495,7 +1478,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="allocator">instance of OrtAllocator</param>
         /// <param name="keys">(output) all keys in the custom metadata map</param>
         /// <param name="numKeys">(output) number of keys in the custom metadata map</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* (OrtStatus*) */ DOrtModelMetadataGetCustomMetadataMapKeys(IntPtr /* (const OrtModelMetadata*) */ modelMetadata,
@@ -1511,7 +1494,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="key">key in the custom metadata map</param>
         /// <param name="value">(output) value for the key in the custom metadata map</param>
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* (OrtStatus*) */ DOrtModelMetadataLookupCustomMetadataMap(IntPtr /* (const OrtModelMetadata*) */ modelMetadata,
@@ -1524,7 +1507,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// Frees ModelMetadata instance
         /// </summary>
         /// <param name="modelMetadata">instance of OrtModelMetadata</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtReleaseModelMetadata(IntPtr /*(OrtModelMetadata*)*/ modelMetadata);
@@ -1535,7 +1518,7 @@ namespace Microsoft.ML.OnnxRuntime
 
         #region Tensor/OnnxValue API
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtGetValue(IntPtr /*(OrtValue*)*/ value,
@@ -1545,35 +1528,35 @@ namespace Microsoft.ML.OnnxRuntime
 
         public static DOrtGetValue OrtGetValue;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtGetValueType(IntPtr /*(OrtValue*)*/ value, out IntPtr /*(OnnxValueType*)*/ onnxtype);
 
         public static DOrtGetValueType OrtGetValueType;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtGetOnnxTypeFromTypeInfo(IntPtr /*(OrtTypeInfo*)*/ typeinfo, out IntPtr /*(OnnxValueType*)*/ onnxtype);
 
         public static DOrtGetOnnxTypeFromTypeInfo OrtGetOnnxTypeFromTypeInfo;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtGetValueCount(IntPtr /*(OrtValue*)*/ value, out IntPtr /*(size_t*)*/ count);
 
         public static DOrtGetValueCount OrtGetValueCount;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtGetTypeInfo(IntPtr /*(OrtValue*)*/ value, IntPtr /*(OrtValue**)*/ typeInfo);
 
         public static DOrtGetTypeInfo OrtGetTypeInfo;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtCreateTensorAsOrtValue(
@@ -1585,7 +1568,7 @@ namespace Microsoft.ML.OnnxRuntime
 
         public static DOrtCreateTensorAsOrtValue OrtCreateTensorAsOrtValue;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* OrtStatus */ DOrtCreateTensorWithDataAsOrtValue(
@@ -1601,7 +1584,7 @@ namespace Microsoft.ML.OnnxRuntime
 
         /// This function doesn't work with string tensor
         /// this is a no-copy method whose pointer is only valid until the backing OrtValue* is free'd.
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtGetTensorMutableData(IntPtr /*(OrtValue*)*/ value, out IntPtr /* (void**)*/ dataBufferHandle);
@@ -1610,7 +1593,7 @@ namespace Microsoft.ML.OnnxRuntime
 
         /// \param value A tensor created from OrtCreateTensor... function.
         /// \param len total data length, not including the trailing '\0' chars.
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtFillStringTensor(
@@ -1620,7 +1603,7 @@ namespace Microsoft.ML.OnnxRuntime
 
         public static DOrtFillStringTensor OrtFillStringTensor;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtGetStringTensorContent(
@@ -1632,7 +1615,7 @@ namespace Microsoft.ML.OnnxRuntime
 
         public static DOrtGetStringTensorContent OrtGetStringTensorContent;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtGetStringTensorDataLength(IntPtr /*(OrtValue*)*/ value,
@@ -1640,7 +1623,7 @@ namespace Microsoft.ML.OnnxRuntime
 
         public static DOrtGetStringTensorDataLength OrtGetStringTensorDataLength;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/
@@ -1648,7 +1631,7 @@ namespace Microsoft.ML.OnnxRuntime
 
         public static DOrtCastTypeInfoToTensorInfo OrtCastTypeInfoToTensorInfo;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtGetTensorTypeAndShape(IntPtr /*(OrtValue*)*/ value, out IntPtr /*(struct OrtTensorTypeAndShapeInfo*)*/ typeAndShapeInfo);
@@ -1656,28 +1639,28 @@ namespace Microsoft.ML.OnnxRuntime
         public static DOrtGetTensorTypeAndShape OrtGetTensorTypeAndShape;
 
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtReleaseTensorTypeAndShapeInfo(IntPtr /*(OrtTensorTypeAndShapeInfo*)*/ value);
 
         public static DOrtReleaseTensorTypeAndShapeInfo OrtReleaseTensorTypeAndShapeInfo;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtGetTensorElementType(IntPtr /*(const struct OrtTensorTypeAndShapeInfo*)*/ typeAndShapeInfo, out IntPtr /*(TensorElementType*)*/ output);
 
         public static DOrtGetTensorElementType OrtGetTensorElementType;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtGetDimensionsCount(IntPtr /*(const struct OrtTensorTypeAndShapeInfo*)*/ typeAndShapeInfo, out UIntPtr output);
 
         public static DOrtGetDimensionsCount OrtGetDimensionsCount;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtGetDimensions(
@@ -1698,7 +1681,7 @@ namespace Microsoft.ML.OnnxRuntime
         *  - the second dimension was an unnamed symbolic dimension (-1 dim value and empty string),
         *  - the entry for the third dimension should be ignored as it is not a symbolic dimension (dim value >= 0).
         */
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtGetSymbolicDimensions(
@@ -1717,14 +1700,14 @@ namespace Microsoft.ML.OnnxRuntime
          * [2,0,4] -> 0
          * [-1,3,4] -> -1
          */
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtGetTensorShapeElementCount(IntPtr /*(const struct OrtTensorTypeAndShapeInfo*)*/ typeAndShapeInfo, out IntPtr /*(long*)*/ output);
 
         public static DOrtGetTensorShapeElementCount OrtGetTensorShapeElementCount;
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtReleaseValue(IntPtr /*(OrtValue*)*/ value);
@@ -1741,7 +1724,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="providers">(output) all execution providers (strings) supported in the native onnxruntime shared library</param>
         /// <param name="numProviders">(output) number of execution providers (strings)</param>
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* (OrtStatus*) */ DOrtGetAvailableProviders(out IntPtr /* (char***) */ providers, out int /* (int*) */ numProviders);
@@ -1754,7 +1737,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// <param name="providers">all execution providers (strings) returned by OrtGetAvailableProviders</param>
         /// <param name="numProviders">number of execution providers (strings)</param>
 
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /* (OrtStatus*) */ DOrtReleaseAvailableProviders(IntPtr /* (char**) */ providers, int /* (int) */ numProviders);
@@ -1765,7 +1748,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// Create an instance of PrepackedWeightsContainer
         /// </summary>
         /// <param name="prepackedWeightsContainer">(output) Created native OrtPrepackedWeightsContainer instance</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate IntPtr /*(OrtStatus*)*/ DOrtCreatePrepackedWeightsContainer(out IntPtr /*(OrtPrepackedWeightsContainer**)*/ prepackedWeightsContainer);
@@ -1776,7 +1759,7 @@ namespace Microsoft.ML.OnnxRuntime
         /// Destroy an instance of PrepackedWeightsContainer
         /// </summary>
         /// <param name="prepackedWeightsContainer">Native OrtPrepackedWeightsContainer instance to be destroyed</param>
-#if __IOS__ || __ANDROID__
+#if __MOBILE__
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 #endif
         public delegate void DOrtReleasePrepackedWeightsContainer(IntPtr /*(OrtPrepackedWeightsContainer*)*/ prepackedWeightsContainer);
