+++ csharp/src/Microsoft.ML.OnnxRuntime/SessionOptions.ios.cs
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace Microsoft.ML.OnnxRuntime
+{
+    /// <summary>
+    /// Represents the platform-specific options for creating an InferenceSession.
+    /// </summary>
+    /// <see cref="https://onnxruntime.ai"/>
+    /// <remarks>Exposes only those APIs supported on iOS.</remarks>
+    public partial class SessionOptions
+    {
+        // TODO: Enable use of CoreML
+    }
+}
