+++ csharp/src/Microsoft.ML.OnnxRuntime/SessionOptions.netstandard.cs
@@ -1,0 +1,20 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.Runtime.InteropServices;
+using System.Text;
+using System.Collections.Generic;
+
+namespace Microsoft.ML.OnnxRuntime
+{
+    /// <summary>
+    /// Represents the platform-specific options for creating an InferenceSession.
+    /// </summary>
+    /// <see cref="https://onnxruntime.ai"/>
+    /// <remarks>Satisfies partial class requirements.</remarks>
+    public partial class SessionOptions
+    {
+
+    }
+}
